pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires      = ["setuptools>=61.0", "wheel"]

[project]
authors = [
    { name = "Zen-Path", email = "zen.path@tuta.com" },
]
dependencies = [
    "pip>=26.0",

    # app
    "Flask==3.1.1",
    "flask-cors==6.0.0",
    "Flask-SQLAlchemy==3.1.1",
    "gallery_dl==1.31.4",
    "bs4==0.0.2",
    "marshmallow==4.2.2",
    "yt-dlp==2026.1.31",
    "Werkzeug==3.1.3",

    # dev
    "pre_commit==4.2.0", # for pre-commit git hooks
    "black==25.1.0",     # for formatting
    "mypy==1.16.1",      # for type checking
    "ruff==0.14.13",     # linter and formatter

    # test
    "pytest==8.4.2",            # for running test suites
    "pytest-playwright==0.7.1", # for UI testing

    # util
    "python-dotenv==1.1.1", # for loading env variables
    "colorama==0.4.6",      # for colored text

    # mypy types
    "types-colorama==0.4.15.20240311",
    "types-Flask-Cors==6.0.0.20250520",
    "types-setuptools==80.9.0.20250529",
    "types-requests==2.32.4.20250913",
]
description = "A script that handles download requests for URLs."
name = "media_server"
readme = "README.md"
requires-python = ">=3.12"
version = "0.0.1"

[tool.setuptools.packages.find]
where = ["app"]

[tool.mypy]
explicit_package_bases = true
ignore_missing_imports = true
strict_optional        = true

[tool.pytest.ini_options]
markers = [
    "ui: checks that require a browser/playwright",
    "slow: marks tests as slow",
]

[tool.ruff]
# Same as Black
indent-width = 4
line-length  = 88

target-version = "py314"

    [tool.ruff.lint]
    # "E" - pycodestyle
    # "F" - pyflakes
    # "I" - isort
    ignore = []
    select = ["E", "F", "I"]

    # Allow fix for all enabled rules (when `--fix`) is provided.
    fixable   = ["ALL"]
    unfixable = []

    [tool.ruff.format]
    # Same as Black
    indent-style              = "space"
    line-ending               = "auto"
    quote-style               = "double"
    skip-magic-trailing-comma = false
